Support TextInput::onBlur
Currently only onBlur event can be supported and the TextInput will be blurred by default onSubmit. This is because there is no way of making ArkUI TextInput stay in edit mode after submitting text and making ArkUI TextArea exit the edit mode after pressing enter. [i.e. blurOnSubmit is always true in singleline and always false in multiline]

Also, `focusable` prop for View was removed, as this caused focus issues in the TextInputs. For the component to be granted focus it and all its ancestors need to have focusable set to true, so we cannot allow setting `focusable: false` in a parent container as not being focusable breaks TextInput behaviour.

related #403

See merge request rnoh/react-native-harmony!522